+++ openBVE/OpenBve/Properties/AssemblyInfo.cs
@@ -1,4 +1,4 @@
-﻿ //1,4,30,2206
+ //1,4,30,2206
  //
  // 
  // This code was generated by a tool. Any changes made manually will be lost
@@ -13,8 +13,8 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyProduct("openBVE")]
 [assembly: AssemblyCopyright("The openBVE Project")]
 [assembly: ComVisible(false)]
-[assembly: AssemblyVersion("1.4.30.635")]
-[assembly: AssemblyFileVersion("1.4.30.635")]
+[assembly: AssemblyVersion("1.4.30.834")]
+[assembly: AssemblyFileVersion("1.4.30.834")]
 [assembly: CLSCompliant(true)]
 
 namespace OpenBve {
